@@ -1,11 +1,9 @@
-from typing import Optional, Tuple
-
 from .consts import KEMId
 from .kem_key import KEMKeyPair
 from .kem_key_interface import KEMKeyInterface
 
 
-class KEMInterface(object):
+class KEMInterface:
     """
     The KEM (Key Encapsulation Mechanism) interface.
     """
@@ -24,11 +22,11 @@ class KEMInterface(object):
         raise NotImplementedError()
 
     def encap(
-        self, pkr: KEMKeyInterface, sks: Optional[KEMKeyInterface] = None, eks: Optional[KEMKeyPair] = None
-    ) -> Tuple[bytes, bytes]:
+        self, pkr: KEMKeyInterface, sks: KEMKeyInterface | None = None, eks: KEMKeyPair | None = None
+    ) -> tuple[bytes, bytes]:
         raise NotImplementedError()
 
-    def decap(self, enc: bytes, skr: KEMKeyInterface, pks: Optional[KEMKeyInterface] = None) -> bytes:
+    def decap(self, enc: bytes, skr: KEMKeyInterface, pks: KEMKeyInterface | None = None) -> bytes:
         raise NotImplementedError()
 
     def derive_key_pair(self, ikm: bytes) -> KEMKeyPair:
