@@ -1,3 +1,5 @@
+from typing import cast
+
 from cryptography.hazmat.primitives.ciphers import aead
 
 from ..aead_key_interface import AEADKeyInterface
@@ -14,7 +16,7 @@ class ChaCha20Poly1305Key(AEADKeyInterface):
         self._ctx = aead.ChaCha20Poly1305(key)
 
     def seal(self, pt: bytes, nonce: bytes, aad: bytes = b"") -> bytes:
-        return self._ctx.encrypt(nonce, pt, aad)
+        return cast(bytes, self._ctx.encrypt(nonce, pt, aad))
 
     def open(self, ct: bytes, nonce: bytes, aad: bytes = b"") -> bytes:
-        return self._ctx.decrypt(nonce, ct, aad)
+        return cast(bytes, self._ctx.decrypt(nonce, ct, aad))
